+++ src/pages/Kubernetes/kubernetesPage.tsx
@@ -20,6 +20,7 @@ import {
   createThanosDatasourceVariable,
 } from '../../variables/datasourceVariables';
 import { copyLinkControl } from '../../scenes/copyLink';
+import { jumpControl } from '../../scenes/objectJump';
 
 const KUBERNETES_ICON = 'public/plugins/debeka-k8s-app/img/kubernetes.png';
 
@@ -73,6 +74,7 @@ export function getKubernetesHomePage() {
       new SceneTimePicker({}),
       new SceneRefreshPicker({ refresh: '1m' }),
       copyLinkControl(),
+      jumpControl(),
     ],
     preserveUrlKeys: ['from', 'to', 'timezone', 'refresh', `var-${THANOS_VARIABLE_NAME}`],
   });